fix some unittest
