+++ work/fixtures.py
@@ -1,3 +1,7 @@
+from core.models import CofkLookupCatalogue
+from work import work_utils
+from work.models import CofkUnionQueryableWork, CofkUnionWork
+
 work_dict_a = dict(
     work_id='work_id value',
     description='description value',
@@ -5,3 +9,13 @@ work_dict_a = dict(
     abstract='abstract value',
     keywords='keywords value',
 )
+
+
+def fixture_queryable_work() -> CofkUnionQueryableWork:
+    work = CofkUnionWork(description='test')
+    work.save()
+
+    q_work = work_utils.clone_queryable_work(work, _return=True)
+    return q_work
+
+
